rm deprecated members file
